refactor: move uploadZipWithFetchApi function to DatasetsUtils
