+++ src/state/sagas/datasets/CreateDataset/CreateDataset.js
@@ -3,32 +3,11 @@
 
 import { call, put, takeEvery } from 'redux-saga/effects';
 import { t } from 'i18next';
-import { Auth } from '@cosmotech/core';
 import { Api } from '../../../../services/config/Api';
 import { DATASET_ACTIONS_KEY } from '../../../commons/DatasetConstants';
 import { dispatchSetApplicationErrorMessage } from '../../../dispatchers/app/ApplicationDispatcher';
 import { INGESTION_STATUS } from '../../../../services/config/ApiConstants';
-
-const uploadZipWithFetchApi = async (organizationId, datasetId, file) => {
-  try {
-    const tokens = await Auth.acquireTokens();
-    const headers = {
-      Authorization: 'Bearer ' + tokens.accessToken,
-      'Content-Type': 'application/octet-stream',
-    };
-    return await fetch(`${Api.defaultBasePath}/organizations/${organizationId}/datasets/${datasetId}`, {
-      method: 'POST',
-      headers,
-      body: file,
-    });
-  } catch (error) {
-    console.error(error);
-    dispatchSetApplicationErrorMessage(
-      error,
-      t('commoncomponents.banner.twingraphNotCreated', 'A problem occurred during twingraph creation')
-    );
-  }
-};
+import { DatasetsUtils } from '../../../../utils/DatasetsUtils';
 
 export function* createDataset(action) {
   const dataset = action.dataset;
@@ -54,7 +33,7 @@ export function* createDataset(action) {
         datasetId: data.id,
       });
     } else if (dataset.sourceType === 'File') {
-      const response = yield call(uploadZipWithFetchApi, organizationId, data.id, dataset.file.file);
+      const response = yield call(DatasetsUtils.uploadZipWithFetchApi, organizationId, data.id, dataset.file.file);
       if (response?.ok) {
         yield put({
           type: DATASET_ACTIONS_KEY.TRIGGER_SAGA_START_TWINGRAPH_STATUS_POLLING,
@@ -73,6 +52,7 @@ export function* createDataset(action) {
     );
   }
 }
+
 function* createDatasetSaga() {
   yield takeEvery(DATASET_ACTIONS_KEY.CREATE_DATASET, createDataset);
 }
