+++ src/utils/DatasetsUtils.js
@@ -1,6 +1,8 @@
 // Copyright (c) Cosmo Tech.
 // Licensed under the MIT license.
 
+import { t } from 'i18next';
+import { Auth } from '@cosmotech/core';
 import {
   CONNECTOR_VERSION_AZURE_STORAGE,
   CONNECTOR_NAME_AZURE_STORAGE,
@@ -8,6 +10,8 @@ import {
   STORAGE_ROOT_DIR_PLACEHOLDER,
 } from '../services/config/ApiConstants';
 import { SecurityUtils } from './SecurityUtils';
+import { Api } from '../services/config/Api';
+import { dispatchSetApplicationErrorMessage } from '../state/dispatchers/app/ApplicationDispatcher';
 
 const patchDatasetWithCurrentUserPermissions = (dataset, userEmail, permissionsMapping) => {
   if (dataset?.security == null || Object.keys(dataset?.security).length === 0) return;
@@ -85,6 +89,27 @@ const removeUndefinedValuesBeforeCreatingDataset = (values) => {
   });
 };
 
+const uploadZipWithFetchApi = async (organizationId, datasetId, file) => {
+  try {
+    const tokens = await Auth.acquireTokens();
+    const headers = {
+      Authorization: 'Bearer ' + tokens.accessToken,
+      'Content-Type': 'application/octet-stream',
+    };
+    return await fetch(`${Api.defaultBasePath}/organizations/${organizationId}/datasets/${datasetId}`, {
+      method: 'POST',
+      headers,
+      body: file,
+    });
+  } catch (error) {
+    console.error(error);
+    dispatchSetApplicationErrorMessage(
+      error,
+      t('commoncomponents.banner.twingraphNotCreated', 'A problem occurred during twingraph creation or update')
+    );
+  }
+};
+
 export const DatasetsUtils = {
   patchDatasetWithCurrentUserPermissions,
   buildStorageFilePath,
@@ -93,4 +118,5 @@ export const DatasetsUtils = {
   buildAzureStorageConnector,
   getAllChildrenDatasetsNames,
   removeUndefinedValuesBeforeCreatingDataset,
+  uploadZipWithFetchApi,
 };
